+++ app/Http/Controllers/MechanicalEmployeeController.php
@@ -6,20 +6,49 @@ use App\Models\MechanicalEmployee;
 use Illuminate\Http\Request;
 use App\Interfaces\EmployeeReadInterface;
 use App\Interfaces\EmployeeWriteInterface;
+use App\Services\EmployeeService;
+use Illuminate\Support\Facades\Gate;
 
-class MechanicalEmployeeController implements EmployeeReadInterface,EmployeeWriteInterface
+class MechanicalEmployeeController extends Controller
 {
+
+    protected $employeeService;
+
+    public function __construct(EmployeeService $employeeService)
+    {
+        $this->employeeService = $employeeService;
+    }
+    
     public function show()
     {
+        if (!Gate::any(['access-admin', 'access-mechanical'])) {
+            abort(403); // Forbidden
+        }
+
         $mechanicalemployees = MechanicalEmployee::all();
         return view('mechanical.home', compact('mechanicalemployees'));
     }
 
     public function create() {
+
+        if (!Gate::any(['access-admin', 'access-mechanical'])) {
+            abort(403); // Forbidden
+        }
+
         return view('mechanical.create');
     }
 
-    public function store(Request $request) {
+    public function store(Request $request)
+    {
+        if (!Gate::any(['access-admin', 'access-mechanical'])) {
+            abort(403); // Forbidden
+        }
+
+        $this->employeeService->store($request, MechanicalEmployee::class);
+        return redirect()->route('mechanical.home')->with('success', 'Mechanical employee added.');
+    }
+
+    /*public function store(Request $request) {
         //dd($request);
         $request->validate([
             'name'=>'required',
@@ -37,15 +66,21 @@ class MechanicalEmployeeController implements EmployeeReadInterface,EmployeeWrit
         $mechanical->save();
         return back()->withSuccess('Employee Added!');
 
-    }
+    }*/
 
     public function shows($id) {
+        if (!Gate::any(['access-admin', 'access-mechanical'])) {
+            abort(403); // Forbidden
+        }
         $mechanicalemployee= MechanicalEmployee::where('id',$id) -> first();
         return view('mechanical.show',['mechanicalemployee'=>$mechanicalemployee]);
     }
 
     public function destroy($id) {
         $mechanicalemployee= MechanicalEmployee::where('id',$id) -> first();
+
+        $this->authorize('delete', $mechanicalemployee);
+
         $mechanicalemployee->delete();
         return back()->withSuccess('Employee deleted!');
     }
